pyproject.toml
[tool.poetry]
name = "testing"
version = "0.1.0"
description = ""
authors = ["Mikhail Fedorov <fedorovm093@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
tqdm = "^4.66.2"
beautifulsoup4 = "^4.12.3"
dockerfile-parse = "^2.0.1"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
